requirements.txt
openai_whisper==20240930
torch==2.7.0+cu128
